Update the docker image
the docker image now includes the depencencies hash as environment
variable DEPENDENCIES_HASH. This allows us to check if the CI is
running inside the correct docker image.
We more metadata is added to the image via the metadata action.

At last, once the PR updating the docker image is merged, the newly
created image is also tagged as `latest` such that it is more convenient
to pull the current dependencies image.