Add solid mechanics tutorial
This tutorial provides a structure, which deforms plastically.
Also, the test is added to the framework-tests.

The general structure of the tutorials chapter is slightly changed,
some more information about the preprocessor cubit has moved to an
introductory section.

The code for generating the reference documentation has been modified,
so that now equations are transformed
from the code to rst whenever a $ sign is encountered.
A registered trademark is also enabled for the documentation.

For some more versatility, the material model gets a new parameter,
TOL, which is given in many other models as well.
Change local Newton tolerance for one test file