Fix bug in cut levelsetinteraction
MPI barriers were inside if statements such that only one process would
wait there, making it wait indefinitely. Though, this code path is
never executed in the tests. The barrier is removed because the function
is called inside a loop over all elements such that the function is not
always run on each process, resulting in a deadlock.

And fix bug where IO::cout was not set up in cut test in case it is used