Rework tensor interpolation framework
- Added error types for tensor interpolation to control
whether the interpolation was successful and (possibly) why not.
- Added struct of parameters used for interpolation.
- Added methods for the computation of the interpolation gradient, i.e.
the derivative of the interpolated matrix with respect to the
interpolation location. Also added a dedicated unit test for it.
- Added dedicated methods for the computation of the material and
spatial stretch tensors.
- Fixed bug in the rotation interpolation: the base matrix is now
correctly determined and decomposed (polar decomposition) before
entering the loop. This fixes the determination of the relative rotation
matrices, where the base matrix decomposition was previously not always
available.
- Added various rotation and eigenvalue interpolation types, as shown in
the publication of Satheesh et al. 10.1002/nme.7373. Only one
combination (R-LOG) is currently implemented.
- Added a more general 1D tensor interpolation unit test, where the
intermediary results of the interpolation algorithm are verified in
detail. Specifically, we look here  at the interpolated eigenvalues
of the logaritmic weighted average routine and the interpolated rotation
vectors.
- Corrected several things in the documentation (e.g. journal names).