Throw if iterative linear solver does not converge
If iterative solvers for linear systems do not converge, the solution is
often not useful at all. To avoid trouble down the road, let's throw
right away, so that one can fix the problem instead of hoping for the
best and being disappointed later on.

If the linear solver does not converge, default behavior is to abort the
simulation. This can be changed and replaced by a warning to the user.