Update geometric search implementation
This PR moves the geometric search out of the FE folder
as it's unrelated to it. In addition the ArborX hash is
updated to reflect the official stable release v2.0.
The input related to geometric search is also moved from
inpar to the respective folder. Some further small fixes
have been done on the way.

Signed-off-by: Max Firmbach <max.firmbach@unibw.de>