Implement SymmetricTensor
This commit introduces the implementation of symmetric tensors in the 4C
linear algebra library. It includes benchmark tests and unit tests.

The symmetric tensor is compatible with the existing tensor library and
also with the stress-like Voigt notation, i.e., one can view a symmetric
tensor as a 6x1 matrix in stress-like Voigt notation (or vice versa).
Additionally, the symmetric fourth-order tensor is compatible to the
linearization of a symmetric second-order tensor in stress-like Voigt
notation w.r.t. a symmetric second-order tensor in strain-like Voigt
notation.